Crate zng_wgt_undo

Undo properties.

Crate

This crate is part of the zng project.

Structs

UndoMix

m Undo scope widget mixin.

Functions

undo_enabled

P Enable or disable undo inside the widget.

undo_interval

P Sets the time interval that undo and redo cover each call for undo handlers in the widget and descendants.

undo_limit

P Sets the maximum length for undo/redo stacks in the widget and descendants.

undo_scope

P Sets if the widget is an undo scope.